feat: #[post_component] registry, manifest, and first real components

+++ app/src/components/counter.rs
@@ -1,7 +1,10 @@
 use leptos::prelude::*;
+use registry::post_component;
 
+/// The v1 interactive demo island: server-rendered, hydrates client-side.
+#[post_component]
 #[island]
-pub fn Counter(initial: i32) -> impl IntoView {
+pub fn Counter(initial: i64) -> impl IntoView {
     let count = RwSignal::new(initial);
 
     view! {
